+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hot-updater/server
 
+## 0.30.10
+
+### Patch Changes
+
+- @hot-updater/core@0.30.10
+- @hot-updater/js@0.30.10
+- @hot-updater/plugin-core@0.30.10
+
 ## 0.30.9
 
 ### Patch Changes